refactor(slack-app): move classify_task_needs_repo into activities/classifiers
Co-locate the Slack-routing classifier with the existing
classify_message_is_agent_directed: both run from the slack_app temporal
worker, both use the same Haiku model and gateway client, both render a
prompt over Slack thread context.

Log identifiers prefixed with slack_app_ to match the rename in
d39cb4c90fb. Tests move from products/slack_app/backend/tests to
posthog/temporal/tests/ai. Activity name is unchanged so workflow replay
is unaffected.

Pure move — no behavior change.
