+++ posthog/temporal/ai/slack_app/eval_slack_repo_selection.py
@@ -81,8 +81,9 @@ from django.conf import settings
 
 from posthog.models import Team
 from posthog.temporal.ai.slack_app import POSTHOG_CODE_SLACK_MENTION_PICKER_GUIDANCE
+from posthog.temporal.ai.slack_app.activities.classifiers import classify_task_needs_repo
 
-from products.slack_app.backend.api import _extract_explicit_repo, classify_task_needs_repo
+from products.slack_app.backend.api import _extract_explicit_repo
 from products.tasks.backend.models import Task
 from products.tasks.backend.repo_selection import (
     RepoSelectionRejectedError,
